# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateScheduleRequest:
    properties:
      available_meals:
        items:
          type: string
        type: array
      date:
        type: string
      day_status:
        type: string
      reason:
        type: string
    required:
    - date
    - day_status
    type: object
...
  handlers.UpdateScheduleRequest:
    properties:
      available_meals:
        items:
          type: string
        type: array
      day_status:
        type: string
      reason:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@craftsbite.com
    name: API Support
  description: CraftsBite Backend API for meal management system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CraftsBite API
  version: "1.0"
paths:
...
  /schedules:
    post:
      consumes:
      - application/json
      description: Create a new meal schedule (Admin/Logistics only)
      parameters:
      - description: Schedule details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Schedule created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create schedule
      tags:
      - schedules
  /schedules/{date}:
    delete:
      description: Delete a meal schedule (Admin/Logistics only)
      parameters:
      - description: Date (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete schedule
      tags:
      - schedules
    get:
      description: Get meal schedule for a specific date
      parameters:
      - description: Date (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get schedule by date
      tags:
      - schedules
    put:
      consumes:
      - application/json
      description: Update an existing meal schedule (Admin/Logistics only)
      parameters:
      - description: Date (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Schedule updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update schedule
      tags:
      - schedules
  /schedules/range:
    get:
      description: Get meal schedules within a date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedules retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get schedule range
      tags:
      - schedules
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"